Added context to downloader and moved method from utils to downloader
